+++ src/index.ts
@@ -1,62 +1,74 @@
 import { onlyExportComponents } from "./only-export-components.ts";
+import type { OnlyExportComponentsOptions } from "./types.d.ts";
 
-export const rules = {
+const rules = {
   "only-export-components": onlyExportComponents,
 };
-
 const plugin = { rules };
 
-export const configs = {
-  recommended: {
-    name: "react-refresh/recommended",
-    plugins: { "react-refresh": plugin },
-    rules: { "react-refresh/only-export-components": "error" },
-  },
-  vite: {
-    name: "react-refresh/vite",
-    plugins: { "react-refresh": plugin },
-    rules: {
-      "react-refresh/only-export-components": [
-        "error",
-        { allowConstantExport: true },
-      ],
-    },
-  },
-  next: {
-    name: "react-refresh/next",
+const buildConfig =
+  ({
+    name,
+    baseOptions,
+  }: {
+    name: string;
+    baseOptions: OnlyExportComponentsOptions;
+  }) =>
+  (options?: OnlyExportComponentsOptions) => ({
+    name: `react-refresh/${name}`,
     plugins: { "react-refresh": plugin },
     rules: {
       "react-refresh/only-export-components": [
         "error",
-        {
-          allowExportNames: [
-            // https://nextjs.org/docs/app/api-reference/file-conventions/route-segment-config
-            "experimental_ppr",
-            "dynamic",
-            "dynamicParams",
-            "revalidate",
-            "fetchCache",
-            "runtime",
-            "preferredRegion",
-            "maxDuration",
-            // https://nextjs.org/docs/app/api-reference/functions/generate-metadata
-            "metadata",
-            "generateMetadata",
-            // https://nextjs.org/docs/app/api-reference/functions/generate-viewport
-            "viewport",
-            "generateViewport",
-            // https://nextjs.org/docs/app/api-reference/functions/generate-image-metadata
-            "generateImageMetadata",
-            // https://nextjs.org/docs/app/api-reference/functions/generate-sitemaps
-            "generateSitemaps",
-            // https://nextjs.org/docs/app/api-reference/functions/generate-static-params
-            "generateStaticParams",
-          ],
-        },
+        { ...baseOptions, ...options },
       ],
     },
-  },
+  });
+
+const configs = {
+  recommended: buildConfig({ name: "recommended", baseOptions: {} }),
+  vite: buildConfig({
+    name: "vite",
+    baseOptions: { allowConstantExport: true },
+  }),
+  next: buildConfig({
+    name: "next",
+    baseOptions: {
+      allowExportNames: [
+        // https://nextjs.org/docs/app/api-reference/file-conventions/route-segment-config
+        "experimental_ppr",
+        "dynamic",
+        "dynamicParams",
+        "revalidate",
+        "fetchCache",
+        "runtime",
+        "preferredRegion",
+        "maxDuration",
+        // https://nextjs.org/docs/app/api-reference/functions/generate-metadata
+        "metadata",
+        "generateMetadata",
+        // https://nextjs.org/docs/app/api-reference/functions/generate-viewport
+        "viewport",
+        "generateViewport",
+        // https://nextjs.org/docs/app/api-reference/functions/generate-image-metadata
+        "generateImageMetadata",
+        // https://nextjs.org/docs/app/api-reference/functions/generate-sitemaps
+        "generateSitemaps",
+        // https://nextjs.org/docs/app/api-reference/functions/generate-static-params
+        "generateStaticParams",
+      ],
+    },
+  }),
 };
 
-// Probably not needed, but keep for backwards compatibility
-export default { rules, configs };
+export const reactRefresh = { plugin, configs };
+
+/** Prefer reactRefresh export which exposed configs as functions */
+export default {
+  rules,
+  configs: {
+    recommended: configs.recommended(),
+    vite: configs.vite(),
+    next: configs.next(),
+  },
+};