+++ src/types.d.ts
@@ -1,15 +1,40 @@
+type Rules = { "only-export-components": any };
+
+export type OnlyExportComponentsOptions = {
+  extraHOCs?: string[];
+  allowExportNames?: string[];
+  allowConstantExport?: boolean;
+  checkJS?: boolean;
+};
+
 type Config = {
-  plugins: { "react-refresh": { rules: Record<string, any> } };
-  rules: Record<string, any>;
+  name: string;
+  plugins: { "react-refresh": { rules: Rules } };
+  rules: Rules;
+};
+type ConfigFn = (options?: OnlyExportComponentsOptions) => {
+  name: string;
+  plugins: { "react-refresh": { rules: Rules } };
+  rules: Rules;
+};
+
+export const reactRefresh: {
+  plugin: {
+    rules: Rules;
+  };
+  configs: {
+    recommended: ConfigFn;
+    vite: ConfigFn;
+    next: ConfigFn;
+  };
 };
 
 declare const _default: {
-  rules: Record<string, any>;
+  rules: Rules;
   configs: {
     recommended: Config;
     vite: Config;
     next: Config;
   };
 };
-
-export = _default;
+export default _default;